goodsbl int to double

+++ Logistics-System/src/bl/goodsbl/Goodsbl.java
@@ -29,8 +29,8 @@ public class Goodsbl {
 		//TODO the parameter of the findByConstName function is to be modified
 		try {
 			//TODO 计算运费
-			int basicprice=ConstBL.findByConstName("").priceConst;
-			int distance=ConstBL.findByConstName("").distanceConst;
+			double basicprice=ConstBL.findByConstName("").priceConst;
+			double distance=ConstBL.findByConstName("").distanceConst;
 			vo.moneyFare=basicprice*distance;//TODO公式先随便写一下
 			vo.moneyTotal=vo.moneyFare+vo.moneyOfPackage;
 		    goodsData.add(vo.toPO(vo));
